fix(typo): match default value to its record field
Hi Jon! I'm assuming this was a tiny typo from when the mmap was created. Before the struct was created, it was a tuple.  Given these are defaults even a `0.` would suffice (right?).

Wanted to clarify re. the above - this was tested locally of course.

P.S. this was one of your best streams yet!